moved tuning into opmodes
